Force float32 in spotify.go

+++ spotify.go
@@ -106,7 +106,7 @@ func (id *ID) String() string {
 // particular artist or playlist.
 type Followers struct {
 	// The total number of followers.
-	Count uint `json:"total"`
+	Count float32 `json:"total"`
 	// A link to the Web API endpoint providing full details of the followers,
 	// or the empty string if this data is not available.
 	Endpoint string `json:"href"`
